+++ examples/12-materialized/go-purego/materialized/metadata/metadata_errors_gen.go
@@ -60,3 +60,6 @@ var ErrLibraryLoad = raw.ErrLibraryLoad
 
 // ErrInvalid represents Zig error.Invalid.
 var ErrInvalid = &Error{Code: 1, Name: "Invalid"}
+
+// ErrOutOfMemory represents Zig error.OutOfMemory.
+var ErrOutOfMemory = &Error{Code: 2, Name: "OutOfMemory"}